Vault Tools

Interact with vault's cryptographic tools. This provides support for high-quality random numbers and cryptographic hashes. This functionality is also available through the transit secret engine.

Methods

random

Generates high-quality random bytes of the specified length. This is totally independent of R's random number stream and provides random numbers suitable for cryptographic purposes.
Usage:
random(bytes = 32, format = "hex")
Arguments:

bytes: Number of bytes to generate (as an integer)

format: The output format to produce; must be one of hex (a single hex string such as d1189e2f83b72ab6), base64 (a single base64 encoded string such as 8TDJekY0mYs=) or raw (a raw vector of length bytes).

hash

Generates a cryptographic hash of given data using the specified algorithm.
Usage:
hash(data, algorithm = NULL, format = "hex")
Arguments:

data: A raw vector of data to hash. To generate a raw vector from an R object, one option is to use unserialize(x, NULL) but be aware that version information may be included. Alternatively, for a string, one might use charToRaw.

algorithm: A string indicating the hash algorithm to use. The exact set of supported algorithms may depend by vault server version, but as of version 1.0.0 vault supports sha2-224, sha2-256, sha2-384 and sha2-512. The default is sha2-256.

format: The format of the output - must be one of hex or base64.

Examples

server <- vaultr::vault_test_server(if_disabled = message)

#> ...waiting for Vault to start

client <- server$client()

# Random bytes in hex
client$tools$random()

#> [1] "414c16dd917861b509d0b6269f698558fbc36b4d0c4bbebf1b9d0d815a076f58"

# base64
client$tools$random(format = "base64")

#> [1] "vmOH3pvxz5watFQLjAGm4wEqgEI4CNpj9a3np3SPKjo="

# raw
client$tools$random(10, format = "raw")

#>  [1] 2e 40 fe bc 4f 93 02 14 a4 f5

# Hash data:
data <- charToRaw("hello vault")
# will produce 55e702...92efd40c2a4
client$tools$hash(data)

#> [1] "55e702c93bd83f5dc1eabdc7e0c268b8a7626b2e8008a7b96023192efd40c2a4"

# sha2-512 hash:
client$tools$hash(data, "sha2-512")

#> [1] "700bfd8ed566cbdcec20ce39db81aec29d489286a97206cd99824d8db1c2e6b3468848766dd791febb1cf7c4dd7faecc98430891698fbe162badfa502186d380"

# cleanup
server$kill()